Force-push everything
I went down this route for several reasons:

1. **It isn't desirable to have a long-lived 'deployment' branch.** It may
be necessary to go back to an earlier or totally different branch on the
remote site, and this is complicated by merging into a long-lived
deployment branch. This branch should act like a tag, showing what is
currently on production, and never cause problems due to merging. So, I
opted to re-create it on every deploy.

2. **We need to be able to overwrite servers.** This is particularly
true of QA servers which may be asked to run many different branches per
day. We need to wipe them clean each time without running into
rejections or conflicts.