Stop deleting the heroku-deploy/env branch during each deploy
I understand the motivations behind [this change](70536ebf049b973c764b0164cddbc6e7d979429c)
but it overlooks the way the plugin uses the heroku-deploy branch to _store precompiled
assets between deploys_. (See delta.rb for how this is/was handled.)

We could work around this by always precompiling assets, but this would leave our version
of the plugin in an inconsistent state. (Do we even need to have the heroku-deploy branch
if we're not using it to store any changes?)

So for now, I'm going to try using the plugin as originally written; we can come up with
a better workaround if needed.

Note: I did put in place an override where the plugin will always precompile when the
public/assets folder is missing. This will help fix scenarios when we force push to the
heroku-deploy/etc branches and the precompiled assets are wiped out.